--- bmi1Conf.txt
@@ -1,2 +1,0 @@
-3650, 2.0, 7.0, 0.2, 0.5
-
