--- bmi2Conf.txt
@@ -1,2 +1,0 @@
-50, 1000, 1000.0, 1
-0.01, 10.0, 0.001
