--- geo.txt
@@ -1,8 +1,0 @@
- 1
-1  6
-0.39	0.07	-0.19	2000000.0	1600000.0	1.05	2.05	0.21
-0.41	0.001	-0.9	2600000.0	2400000.0	0.812	2.03	0.15
-0.38	0.06	-0.6	2600000.0	2400000.0	1.21	2.13	0.60
-0.35	0.06	-0.324	2900000.0	2000000.0	1.42	2.52	7.04
-0.28	0.018	-0.109	3100000.0	2000000.0	1.78	2.04	17.0
-0.05	0.067	-0.215	3000000.0	2500000.0	2.45	2.62	8.0
