--- grid.txt
@@ -1,190 +1,0 @@
-176
--1.5
--1.45
--1.4
--1.35
--1.3
--1.25
--1.2
--1.15
--1.1
--1.05
--1
--0.95
--0.9
--0.85
--0.8
--0.75
--0.7
--0.65
--0.6
--0.55
--0.5
--0.45
--0.4
--0.35
--0.3
--0.25
--0.2
--0.15
--0.1
--0.08
--0.06
--0.04
--0.03
--0.02
--0.01
--0.005
--0.002
--0.001
-0
-0.001
-0.002
-0.005
-0.01
-0.02
-0.03
-0.04
-0.05
-0.08
-0.1
-0.12
-0.14
-0.16
-0.18
-0.2
-0.22
-0.24
-0.26
-0.28
-0.3
-0.32
-0.34
-0.36
-0.38
-0.4
-0.42
-0.44
-0.46
-0.48
-0.5
-0.52
-0.54
-0.56
-0.58
-0.6
-0.62
-0.64
-0.66
-0.68
-0.7
-0.72
-0.74
-0.76
-0.78
-0.8
-0.82
-0.84
-0.86
-0.88
-0.9
-0.92
-0.94
-0.96
-0.98
-1
-1.05
-1.1
-1.15
-1.2
-1.25
-1.3
-1.35
-1.4
-1.45
-1.5
-1.55
-1.6
-1.65
-1.7
-1.75
-1.8
-1.85
-1.9
-1.95
-2
-2.1
-2.2
-2.3
-2.4
-2.5
-2.6
-2.7
-2.8
-2.9
-3
-3.1
-3.2
-3.3
-3.4
-3.5
-3.6
-3.7
-3.8
-3.9
-4
-4.1
-4.2
-4.3
-4.4
-4.5
-4.6
-4.7
-4.8
-5
-5.3
-5.7
-6
-6.5
-7
-7.5
-8
-8.5
-9
-9.5
-10
-10.5
-11
-11.5
-12
-12.5
-13
-14
-15
-16
-17
-18
-19
-20
-21
-23
-28
-33
-50
-60
-70
-80
-90
-12
-40
-48
-50
-54
-58
-62
-65
-69
-73
-81
-88
-96
