--- initial.txt
@@ -1,15 +1,0 @@
-    1   13
- DEPTH  TEMP
--1.5	2.861
-0.001	3.992
-0.072	3.433
-0.125	2.591
-0.2	1.094
-0.277	0.001
-0.354	-0.264
-0.424	-0.589
-0.506	-0.847
-0.583	-1.114
-0.741	-1.871
-0.885	-2.346
-1.1	-2.945
