--- mergedConf.txt
@@ -1,6 +1,0 @@
-3650, 2.0, 7.0, 0.2, 0.5
-
-Start of config file 2: 
-
-50, 1000, 1000.0, 1
-0.01, 10.0, 0.001
