--- snow.txt
@@ -1,366 +1,0 @@
-365
-1     0.000
-2	   0.000
-3	   0.000
-4	   0.000
-5	   0.000
-6	   0.000
-7	   0.000
-8	   0.000
-9	   0.000
-10	   0.000
-11	   0.000
-12	   0.000
-13	   0.000
-14	   0.000
-15	   0.000
-16	   0.000
-17	   0.000
-18	   0.000
-19	   0.000
-20	   0.000
-21	   0.000
-22	   0.000
-23	   0.000
-24	   0.000
-25	   0.000
-26	   0.000
-27	   0.000
-28	   0.000
-29	   0.000
-30	   0.000
-31	   0.000
-32	   0.000
-33	   0.000
-34	   0.000
-35	   0.000
-36	   0.000
-37	   0.000
-38	   0.000
-39	   0.000
-40	   0.000
-41	   0.000
-42	   0.000
-43	   0.000
-44	   0.000
-45	   0.000
-46	   0.000
-47	   0.000
-48	   0.000
-49	   0.000
-50	   0.000
-51	   0.000
-52	   0.000
-53	   0.000
-54	   0.000
-55	   0.000
-56	   0.000
-57	   0.000
-58	   0.000
-59	   0.000
-60	   0.000
-61	   0.000
-62	   0.000
-63	   0.000
-64	   0.000
-65	   0.000
-66	   0.000
-67	   0.000
-68	   0.000
-69	   0.000
-70	   0.000
-71	   0.000
-72	   0.000
-73	   0.000
-74	   0.000
-75	   0.000
-76	   0.000
-77	   0.000
-78	   0.000
-79	   0.000
-80	   0.000
-81	   0.000
-82	   0.000
-83	   0.000
-84	   0.000
-85	   0.000
-86	   0.000
-87	   0.000
-88	   0.000
-89	   0.000
-90	   0.010
-91	   0.011
-92	   0.012
-93	   0.013
-94	   0.014
-95	   0.015
-96	   0.016
-97	   0.017
-98	   0.018
-99	   0.019
-100	   0.020
-101	   0.021
-102	   0.022
-103	   0.023
-104	   0.024
-105	   0.026
-106	   0.027
-107	   0.028
-108	   0.028
-109	   0.029
-110	   0.030
-111	   0.031
-112	   0.032
-113	   0.033
-114	   0.034
-115	   0.035
-116	   0.036
-117	   0.037
-118	   0.037
-119	   0.038
-120	   0.039
-121	   0.040
-122	   0.040
-123	   0.041
-124	   0.042
-125	   0.042
-126	   0.043
-127	   0.043
-128	   0.044
-129	   0.044
-130	   0.044
-131	   0.044
-132	   0.045
-133	   0.045
-134	   0.045
-135	   0.045
-136	   0.045
-137	   0.045
-138	   0.044
-139	   0.044
-140	   0.044
-141	   0.044
-142	   0.044
-143	   0.043
-144	   0.043
-145	   0.043
-146	   0.042
-147	   0.042
-148	   0.042
-149	   0.041
-150	   0.041
-151	   0.041
-152	   0.040
-153	   0.040
-154	   0.040
-155	   0.040
-156	   0.040
-157	   0.039
-158	   0.039
-159	   0.039
-160	   0.040
-161	   0.040
-162	   0.040
-163	   0.040
-164	   0.040
-165	   0.040
-166	   0.041
-167	   0.041
-168	   0.042
-169	   0.042
-170	   0.042
-171	   0.043
-172	   0.044
-173	   0.044
-174	   0.045
-175	   0.046
-176	   0.047
-177	   0.048
-178	   0.049
-179	   0.050
-180	   0.051
-181	   0.052
-182	   0.053
-183	   0.054
-184	   0.055
-185	   0.056
-186	   0.057
-187	   0.058
-188	   0.060
-189	   0.061
-190	   0.062
-191	   0.062
-192	   0.063
-193	   0.064
-194	   0.065
-195	   0.065
-196	   0.066
-197	   0.066
-198	   0.066
-199	   0.067
-200	   0.067
-201	   0.067
-202	   0.067
-203	   0.067
-204	   0.067
-205	   0.068
-206	   0.068
-207	   0.068
-208	   0.069
-209	   0.069
-210	   0.070
-211	   0.071
-212	   0.072
-213	   0.073
-214	   0.075
-215	   0.077
-216	   0.079
-217	   0.081
-218	   0.083
-219	   0.086
-220	   0.088
-221	   0.091
-222	   0.094
-223	   0.097
-224	   0.101
-225	   0.104
-226	   0.108
-227	   0.111
-228	   0.115
-229	   0.119
-230	   0.123
-231	   0.127
-232	   0.130
-233	   0.134
-234	   0.138
-235	   0.142
-236	   0.146
-237	   0.149
-238	   0.153
-239	   0.157
-240	   0.160
-241	   0.164
-242	   0.167
-243	   0.170
-244	   0.173
-245	   0.176
-246	   0.179
-247	   0.181
-248	   0.183
-249	   0.185
-250	   0.187
-251	   0.188
-252	   0.190
-253	   0.191
-254	   0.191
-255	   0.192
-256	   0.192
-257	   0.193
-258	   0.193
-259	   0.193
-260	   0.193
-261	   0.192
-262	   0.192
-263	   0.191
-264	   0.190
-265	   0.189
-266	   0.188
-267	   0.187
-268	   0.185
-269	   0.184
-270	   0.183
-271	   0.182
-272	   0.181
-273	   0.180
-274	   0.179
-275	   0.179
-276	   0.179
-277	   0.180
-278	   0.180
-279	   0.182
-280	   0.183
-281	   0.185
-282	   0.188
-283	   0.191
-284	   0.194
-285	   0.198
-286	   0.202
-287	   0.206
-288	   0.211
-289	   0.217
-290	   0.222
-291	   0.228
-292	   0.235
-293	   0.241
-294	   0.249
-295	   0.256
-296	   0.263
-297	   0.271
-298	   0.279
-299	   0.286
-300	   0.294
-301	   0.301
-302	   0.308
-303	   0.314
-304	   0.321
-305	   0.326
-306	   0.331
-307	   0.335
-308	   0.338
-309	   0.341
-310	   0.343
-311	   0.343
-312	   0.343
-313	   0.342
-314	   0.340
-315	   0.337
-316	   0.333
-317	   0.329
-318	   0.323
-319	   0.316
-320	   0.309
-321	   0.300
-322	   0.291
-323	   0.281
-324	   0.271
-325	   0.260
-326	   0.248
-327	   0.236
-328	   0.224
-329	   0.211
-330	   0.197
-331	   0.184
-332	   0.121
-333	   0.05
-334	   0.000
-335	   0.000
-336	   0.000
-337	   0.000
-338	   0.000
-339	   0.000
-340	   0.000
-341	   0.000
-342	   0.000
-343	   0.000
-344	   0.000
-345	   0.000
-346	   0.000
-347	   0.000
-348	   0.000
-349	   0.000
-350	   0.000
-351	   0.000
-352	   0.000
-353	   0.000
-354	   0.000
-355	   0.000
-356	   0.000
-357	   0.000
-358	   0.000
-359	   0.000
-360	   0.000
-361	   0.000
-362	   0.000
-363	   0.000
-364	   0.000
-365	   0.000
