--- test/bmi1Conf.txt
@@ -1,19 +1,0 @@
-Snow Class[SNOW CLIMATE CLASS, STURM ET AL 1995 (CODE FROM 0-7)]
-1
-Open Area or Not [1:Open, 0:forested] FOREST ZONE (MOST SNOW DEPTHS MEASURED AT OPEN SITES)
-1
-Adjust factor of precipitation [decimal]
-1.0
-Initial Snow Depth [unit:cm]
-0.0
-Initial Snow Density [unit:kg/m3]
-200.0
-Total Time Steps:
-60
-Input Files Path
-
-Land Surface Air Temperature [unit:degree C], [StName: land_surface_air__temperature]
-tair.csv
-Total precipitation in a time step [unit:mm], [StName: precipitation_mass_flux]
-prec.csv
-
