--- test/bound.txt
@@ -1,366 +1,0 @@
-365
-1  2.861
-2	2.633
-3	4.417
-4	3.715
-5	4.884
-6	7.242
-7	11.392
-8	8.799
-9	10.312
-10	10.063
-11	7.578
-12	5.289
-13	4.694
-14	6.013
-15	8.297
-16	12.909
-17	10.614
-18	7.266
-19	10.3
-20	14.891
-21	10.42
-22	6.722
-23	5.862
-24	5.915
-25	4.952
-26	6.248
-27	8.157
-28	7.741
-29	9.25
-30	7.398
-31	7.283
-32	8.344
-33	9.153
-34	5.683
-35	6.576
-36	7.128
-37	7.524
-38	7.505
-39	6.193
-40	4.214
-41	4.108
-42	4.42
-43	5.578
-44	7.315
-45	12.505
-46	13.848
-47	13.288
-48	12.362
-49	10.019
-50	9.776
-51	8.137
-52	5.812
-53	4.303
-54	4.082
-55	9.054
-56	11.002
-57	5.584
-58	5.099
-59	4.623
-60	4.863
-61	6.392
-62	4.398
-63	4.258
-64	4.434
-65	2.428
-66	2.273
-67	3.68
-68	4.906
-69	4.289
-70	3.04
-71	3.396
-72	4.317
-73	3.219
-74	8.118
-75	10.4
-76	9.139
-77	4.111
-78	6.728
-79	5.428
-80	2.463
-81	2.403
-82	3.456
-83	2.816
-84	1.229
-85	0.979
-86	-0.098
-87	-0.054
-88	-1.505
-89	-0.378
-90	-0.43
-91	-0.195
-92	-1.029
-93	-1.766
-94	-4.354
-95	-5.802
-96	-2.488
-97	-2.407
-98	-5.439
-99	-7.492
-100	-6.772
-101	-9.483
-102	-10.102
-103	-11.825
-104	-10.598
-105	-3.911
-106	-5.459
-107	-4.817
-108	-4.606
-109	-3.748
-110	-4.638
-111	-7.505
-112	-3.63
-113	-4.776
-114	-2.86
-115	-6.297
-116	-7.748
-117	-10.589
-118	-12.18
-119	-15.586
-120	-13.391
-121	-8.811
-122	-9.236
-123	-8.923
-124	-7.14
-125	-6.745
-126	-12.382
-127	-10.452
-128	-6.836
-129	-8.37
-130	-10.59
-131	-10.989
-132	-12.187
-133	-15.257
-134	-15.718
-135	-8.684
-136	-12.778
-137	-11.76
-138	-9.467
-139	-11.793
-140	-8.326
-141	-10.992
-142	-12.107
-143	-13.508
-144	-15.798
-145	-13.549
-146	-13.475
-147	-7.814
-148	-8.751
-149	-6.644
-150	-8.957
-151	-14.609
-152	-17.804
-153	-24.09
-154	-23.79
-155	-17.815
-156	-10.45
-157	-8.104
-158	-9.105
-159	-12.195
-160	-21.449
-161	-24.464
-162	-18.745
-163	-8.2
-164	-12.407
-165	-16.478
-166	-13.675
-167	-14.015
-168	-11.805
-169	-15.369
-170	-15.137
-171	-21.705
-172	-21.997
-173	-29.205
-174	-35.295
-175	-35.391
-176	-38.902
-177	-38.77
-178	-29.075
-179	-28.805
-180	-23.775
-181	-21.853
-182	-20.045
-183	-24.702
-184	-17.629
-185	-17.274
-186	-19.367
-187	-20.435
-188	-21.959
-189	-18.577
-190	-19.078
-191	-26.678
-192	-32.267
-193	-36.295
-194	-36.547
-195	-38.664
-196	-37.63
-197	-35.735
-198	-35.489
-199	-42.228
-200	-44.585
-201	-41.17
-202	-40.697
-203	-34.57
-204	-26.853
-205	-29.757
-206	-22.361
-207	-2.526
-208	-13.738
-209	-10.035
-210	-19.152
-211	-30.232
-212	-37.822
-213	-31.569
-214	-34.773
-215	-27.483
-216	-30.676
-217	-34.878
-218	-29.201
-219	-29.165
-220	-32.716
-221	-34.429
-222	-32.728
-223	-26.78
-224	-38.515
-225	-38.543
-226	-39.035
-227	-36.104
-228	-34.723
-229	-29.808
-230	-30.495
-231	-27.77
-232	-23.914
-233	-28.978
-234	-25.394
-235	-24.283
-236	-22.271
-237	-20.464
-238	-10.462
-239	-10.661
-240	-13.396
-241	-24.097
-242	-28.241
-243	-36.334
-244	-36.201
-245	-24.963
-246	-28.853
-247	-39.232
-248	-39.951
-249	-35.069
-250	-34.377
-251	-29.352
-252	-21.569
-253	-28.608
-254	-26.216
-255	-24.641
-256	-22.705
-257	-21.873
-258	-25.324
-259	-31.352
-260	-32.903
-261	-34.715
-262	-32.05
-263	-29.99
-264	-31.965
-265	-40.98
-266	-41.448
-267	-39.988
-268	-37.641
-269	-35.223
-270	-36.533
-271	-33.395
-272	-30.095
-273	-33.31
-274	-29.762
-275	-27.895
-276	-27.36
-277	-21.966
-278	-12.797
-279	-12.596
-280	-6.76
-281	-11.142
-282	-10.48
-283	-8.31
-284	-10.943
-285	-9.79
-286	-13.697
-287	-18.439
-288	-22.163
-289	-25.235
-290	-24.938
-291	-19.432
-292	-16.782
-293	-17.988
-294	-20.288
-295	-20.769
-296	-14.705
-297	-2.864
-298	-7.974
-299	-9.161
-300	-8.952
-301	-9.251
-302	-5.337
-303	-5.422
-304	-10.901
-305	-7.227
-306	-10.978
-307	-12.303
-308	-12.365
-309	-9.689
-310	-9.683
-311	-8.84
-312	-6.048
-313	-5.948
-314	-5.589
-315	-5.557
-316	-6.677
-317	-6.844
-318	-5.156
-319	-8.531
-320	-7.085
-321	-4.574
-322	-3.114
-323	-2.277
-324	-1.871
-325	-3.474
-326	-3.656
-327	-3.094
-328	-3.344
-329	-2.668
-330	-0.795
-331	2.453
-332	3.933
-333	2.32
-334	4.165
-335	4.008
-336	0.736
-337	2.55
-338	1.198
-339	-0.651
-340	-0.88
-341	1.318
-342	2.551
-343	6.473
-344	3.58
-345	4.215
-346	3.583
-347	2.324
-348	1.486
-349	3.992
-350	5.341
-351	5.157
-352	5.666
-353	7.667
-354	10.441
-355	6.029
-356	7.97
-357	4.829
-358	7.193
-359	11.072
-360	11.95
-361	6.985
-362	11.38
-363	7.983
-364	11.033
-365	11.793
