--- test/mergedConf.txt
@@ -1,57 +1,0 @@
-Snow Class[SNOW CLIMATE CLASS, STURM ET AL 1995 (CODE FROM 0-7)]
-1
-Open Area or Not [1:Open, 0:forested] FOREST ZONE (MOST SNOW DEPTHS MEASURED AT OPEN SITES)
-1
-Adjust factor of precipitation [decimal]
-1.0
-Initial Snow Depth [unit:cm]
-0.0
-Initial Snow Density [unit:kg/m3]
-200.0
-Total Time Steps:
-60
-Input Files Path
-
-Land Surface Air Temperature [unit:degree C], [StName: land_surface_air__temperature]
-tair.csv
-Total precipitation in a time step [unit:mm], [StName: precipitation_mass_flux]
-prec.csv
-
-Start of config file 2: 
-
-input files
-sites.txt
-bound.txt
-snow.txt
-rsnow.txt
-initial.txt
-grid.txt
-vegetation.txt
-geo.txt
-
-output files
-mean.txt
-result.txt
-start.txt
-
-restart: 0: start from previous time step / 1: start from the beginning
- 1
-step | taum | tmin :
- 1.0   0.1   0.001
-begin end : start and end, in the example it runs over one year from 0 to 1
- 0.0   1.0
-smoothing factor | unfrozen water parameter | max number of iterations
- 0.01   0.01   5
-number of second in a day [sec] | number of time steps (in the example number of days in a year )
- 86400.0   365
-sea level | max number of freezing fronts [integer number]
- 0.0   4
-freezing front min and max depth [meters]      
- 0.05   10.0
-saturation coefficient (fraction of 1)   
- 0.95
-input_folder [leave the follow line as blank, if you want to use the same dir with this cfg file]
-
-output_folder [leave the follow line as blank, if you want to use the same dir with this cfg file] 
-
-end configuration
