--- test/sites.txt
@@ -1,3 +1,0 @@
- 1
- 
-          246   1	1	1	1	0.00
