--- test/vegetation.txt
@@ -1,2 +1,0 @@
- 1
- 1  0
